// src/api-clients/mealie/types.gen.ts
export type IngredientUnitOutput = {
    id: string;
    name: string;
    pluralName?: string | null;
    description?: string;
    extras?: {
        [key: string]: unknown;
    } | null;
    fraction?: boolean;
    abbreviation?: string;
    pluralAbbreviation?: string | null;
    useAbbreviation?: boolean;
    aliases?: Array<IngredientUnitAlias>;
    createdAt?: string | null;
    updatedAt?: string | null;
};

export type IngredientUnitAlias = {
    name: string;
};